model: drop cumulative_success_rate and 2 more features
This project is more about what we can do rather than what can be done.
it is not viable for me to track cumulative_success_rate, num_mistakes, concept_count without having dual sync support from both backend as well as chrome extension.
The amount of work that can be put into this is infinite, but there's only so much i can do. Maybe in future I will consider it. For now, dropping these 3 features is the best option.

Signed-off-by: ca5 <catinice@outlook.com>
Change-Id: Ia463103169b0f65f91fc6a6b4a33b5037eabc9f2